Skip remote commits status on orphan branch.
Use merge-base to detect a common ancestor between
the remote tracking branch and origin/master.
If no common ancestor is found the remote commits
status is not displayed because it is assumed the branches
have no common history.
This is the case for orphan branches which are completely
separate from origin/master like the popular gh-pages branch
created when setting up a github pages branch.